Added AutoMapper, and two basic mappers
- Also added a controller that can be used to manually test mappings

+++ backend/StrategyGame.Bll/DTO/Sent/Country/CountryInfo.cs
@@ -13,8 +13,8 @@ namespace StrategyGame.Bll.Dto.Sent.Country
         public int Round { get; set; }
         public int Rank { get; set; }
         public IEnumerable<UnitInfo> ArmyInfo { get; set; }
-        public int Pearls { get; set; }
-        public int Corals { get; set; }
+        public long Pearls { get; set; }
+        public long Corals { get; set; }
         public BriefCreationInfo Buildings { get; set; }
         public BriefCreationInfo Researches { get; set; }
     }
